requirements.txt
bcrypt
cryptography
fastapi
python-jose
python-multipart
sqlalchemy
uvicorn
passlib
alembic